fix: category NavLink 사용해서 구현
