+++ src/components/common/FlexButton.jsx
@@ -1,3 +1,4 @@
+/** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 
 const defaultProps = {
