--- src/components/shop/rawData.js
@@ -1,62 +1,0 @@
-export const items = [
-	{
-		id: 1,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 2,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 3,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 4,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 5,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 6,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 7,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 8,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 9,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-	{
-		id: 10,
-		img_url: 'https://cdn.imweb.me/upload/S201905295cee7c0f94cee/12d4d58e92dd7.jpeg',
-		name: '이름',
-		price: 100000,
-	},
-];
